testing/web-platform/tests/shadow-dom/declarative/getinnerhtml.tentative.html

<!DOCTYPE html>
<title>getInnerHTML </title>
<link rel='author' title='Mason Freed' href='mailto:masonfreed@chromium.org'>
<link rel='help' href='https://github.com/whatwg/dom/issues/831'>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='../resources/shadow-dom-utils.js'></script>

<body>

<script>
function testElementType(allowsShadowDom, elementType, mode, delegatesFocus) {
  var t = test(t => {
    // Create and attach element
    const wrapper = document.createElement('div');
    t.add_cleanup(function() { wrapper.remove(); });
    document.body.appendChild(wrapper);
    const element = document.createElement(elementType);
    wrapper.appendChild(element);

    const isOpen = mode === 'open';
    if (allowsShadowDom) {
      let correctShadowHtml = isOpen ? `<template shadowroot="${mode}"><slot></slot></template>` : '';
      let correctHtml = `<${elementType}>${correctShadowHtml}</${elementType}>`;
      const shadowRoot = element.attachShadow({mode: mode, delegatesFocus: delegatesFocus});
      shadowRoot.appendChild(document.createElement('slot'));
      assert_equals(wrapper.getInnerHTML(true),correctHtml);
    } else {
      // For non-shadow hosts, getInnerHTML(true) should also match .innerHTML
      assert_equals(wrapper.getInnerHTML(true),wrapper.innerHTML);
    }

    // Either way, make sure getInnerHTML() matches .innerHTML
    assert_equals(wrapper.getInnerHTML(false),wrapper.innerHTML,'getInnerHTML() with includeShadowRoots false should return the same as .innerHTML');
    assert_equals(wrapper.getInnerHTML(),wrapper.innerHTML,'getInnerHTML() with no arguments should return the same as .innerHTML');

  }, `getInnerHTML() on <${elementType}>${allowsShadowDom ? `, with mode=${mode}, delegatesFocus=${delegatesFocus}.` : ''}`);
}

function runAllTests() {
  const allElements = HTML5_ELEMENT_NAMES;
  const safelisted = ATTACHSHADOW_SAFELISTED_ELEMENTS;
  const minimumKnownElements = 107; // We should have at least this many elements in the lists from shadow-dom-utils.js.
  assert_true(allElements.length >= minimumKnownElements,'All element types should be tested');
  for (let elementName of allElements) {
    const allowsShadowDom = safelisted.includes(elementName);
    if (allowsShadowDom) {
      for (let delegatesFocus of [false, true]) {
        for (let mode of ['open', 'closed']) {
          testElementType(true, elementName, mode, delegatesFocus);
        }
      }
    } else {
      testElementType(false, elementName);
    }
  }
}

runAllTests();

</script>
